+++ libft/srcs/ft_memset.c
@@ -1,4 +1,6 @@
-void	*ft_memset(void *memptr, int val, int num)
+#include "ft.h"
+
+void	*ft_memset(void *memptr, int val, size_t num)
 {
 	void *buf;
